parser/requirements.txt
beautifulsoup4==4.7.1
lxml==4.3.3
psycopg2==2.7.7
requests==2.21.0
SQLAlchemy==1.3.1
urllib3==1.24.1
